scripts/run_merge.py
import sys
import os
import yaml

sys.path.append(os.path.join(os.path.dirname(__file__), "..", "src"))
from uzombie.modeling.merging import dare_ties_merge

def main():
    # Example usage: python scripts/run_merge.py --models path1 path2 --base meta-llama/Llama-3-8b
    # Simplified for now: reads 'merge.yaml' if exists or uses args
    # Here we just demo the call
    pass 
    # Logic to parse args and call dare_ties_merge goes here
    # (Left brief as user request focused on structure)

if __name__ == "__main__":
    main()