requirements.txt
absl-py==2.3.1
accelerate==1.12.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
bitsandbytes==0.49.0
certifi==2025.11.12
chardet==5.2.0
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6
cut-cross-entropy==25.1.1
dataproperty==1.1.0
datasets==4.3.0
diffusers==0.36.0
dill==0.4.0
docstring-parser==0.17.0
evaluate==0.4.6
filelock==3.20.1
frozenlist==1.8.0
fsspec==2025.9.0
h11==0.16.0
hf-transfer==0.1.9
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
immutables==0.21
importlib-metadata==8.7.1
jinja2==3.1.6
joblib==1.5.3
jsonlines==4.0.0
liger-kernel==0.6.4
lm-eval==0.4.9.2
lxml==6.0.2
markdown-it-py==4.0.0
markupsafe==3.0.3
mbstrdecoder==1.1.4
mdurl==0.1.2
mergekit==0.1.4
more-itertools==10.8.0
mpmath==1.3.0
msgspec==0.20.0
multidict==6.7.0
multiprocess==0.70.16
networkx==3.6.1
nltk==3.9.2
numexpr==2.14.1
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pathvalidate==3.3.1
peft==0.18.0
pillow==12.0.0
portalocker==3.2.0
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
pyarrow==22.0.0
pybind11==3.0.1
pydantic==2.10.6
pydantic-core==2.27.2
pygments==2.19.2
pytablewriter==1.2.1
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rouge-score==0.1.2
sacrebleu==2.5.1
safetensors==0.5.3
scikit-learn==1.8.0
scipy==1.16.3
sentencepiece==0.2.1
six==1.17.0
sqlitedict==2.1.0
sympy==1.13.1
tabledata==1.3.4
tabulate==0.9.0
tcolorpy==0.1.7
tensorly==0.9.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.5.0
torchaudio==2.5.0+cu124
torchvision==0.20.0
tqdm==4.67.1
tqdm-multiprocess==0.0.11
transformers==4.57.3
triton==3.1.0
trl==0.24.0
typeguard==4.4.4
typepy==1.3.4
typing-extensions==4.15.0
tyro==1.0.3
tzdata==2025.3
unsloth==2025.12.9
unsloth-zoo==2025.12.7
urllib3==2.6.2
uzombie==1.0.2
word2number==1.1
xformers==0.0.28.post2
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0

# Editable install for local development (adjust path if needed)
-e .